Turned on vsync and changed misleading varible type, motion fails without vsync

+++ src/mintSDL/timer.h
@@ -6,7 +6,7 @@ struct MintTimer {
 	char isPaused;
 	int ticks;
 	int frames;
-	float elapsed;
+	int elapsed;
 	float msPerReport;
 	float msSinceLastReport;
 };
